# Fix KafkaConsumerActor excessive CPU consumption