Fixes #270

## Changes

Moved `Confluent.Kafka.AdminClient` instantiation closer to where it needed to be instead of caching an instance of it.

The problematic thread is gone:
![image](https://user-images.githubusercontent.com/8293237/151072676-3d97070a-21ae-41fc-851d-c3c40e9d2372.png)