@Aaronontheweb 2022-01-26
From the numbers you posted in Slack:

## Before

CPU Benchmark complete.

BeforeLoad
 CPU Usage Mode | Mean | StdErr | StdDev | Median | Maximum |
--------------- |----- |------- |------- |------- |-------- |
 User | 4,954.688 ms | 12.780 ms | 40.414 ms | 4,968.750 ms | 5,000.000 ms |
 Kernel | 9.375 ms | 2.421 ms | 7.655 ms | 15.625 ms | 15.625 ms |
 Total | 4,964.062 ms | 13.082 ms | 41.369 ms | 4,984.375 ms | 5,015.625 ms |


AfterLoad
 CPU Usage Mode | Mean | StdErr | StdDev | Median | Maximum |
--------------- |----- |------- |------- |------- |-------- |
 User | 4,948.438 ms | 15.157 ms | 47.931 ms | 4,968.750 ms | 5,000.000 ms |
 Kernel | 9.375 ms | 3.953 ms | 12.500 ms | 0.000 ms | 31.250 ms |
 Total | 4,957.812 ms | 14.498 ms | 45.848 ms | 4,968.750 ms | 5,000.000 ms |

## After

CPU Benchmark complete.

BeforeLoad
 CPU Usage Mode | Mean | StdErr | StdDev | Median | Maximum |
--------------- |----- |------- |------- |------- |-------- |
 User | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms |
 Kernel | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms |
 Total | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms |


AfterLoad
 CPU Usage Mode | Mean | StdErr | StdDev | Median | Maximum |
--------------- |----- |------- |------- |------- |-------- |
 User | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms |
 Kernel | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms |
 Total | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms | 0.000 ms |

This means that the idle CPU was essentially not measurable.